refactor to bring grouping code into a separate file
