+++ enclone_tail/src/lib.rs
@@ -4,6 +4,7 @@ extern crate enclone_core;
 
 pub mod display_tree;
 pub mod group;
+pub mod grouper;
 pub mod neighbor;
 pub mod newick;
 pub mod plot;
